add data layer for update widget

+++ src/Han.Server.Business/Commands/Widgets/UpdateWidget/IUpdateWidgetCommand.cs
@@ -2,5 +2,5 @@ using Han.Server.Business.Models.Widgets.UpdateWidget;
 
 namespace Han.Server.Business.Commands.Widgets.UpdateWidget
 {
-    public interface IUpdateWidgetCommand : ICommand<UpdateWidgetCommandRequestModel, byte> { }
+    public interface IUpdateWidgetCommand : ICommand<UpdateWidgetCommandRequestModel, ulong> { }
 }
